--- a/base32.go
+++ b/base32.go
@@ -82,41 +82,31 @@ func (enc *Encoding) Encode(dst, src []byte) {
 
 	// Add the remaining small block
 	if len(src) > 0 {
-		var b [8]byte
-
-		// Unpack 8x 5-bit source blocks into a 5 byte
-		// destination quantum
+		var val uint64
 		switch len(src) {
 		default:
-			b[7] = src[4] & 0x1F
-			b[6] = src[4] >> 5
+			val |= uint64(src[4])
 			fallthrough
 		case 4:
-			b[6] |= (src[3] << 3) & 0x1F
-			b[5] = (src[3] >> 2) & 0x1F
-			b[4] = src[3] >> 7
+			val |= uint64(src[3]) << 8
 			fallthrough
 		case 3:
-			b[4] |= (src[2] << 1) & 0x1F
-			b[3] = (src[2] >> 4) & 0x1F
+			val |= uint64(src[2]) << 16
 			fallthrough
 		case 2:
-			b[3] |= (src[1] << 4) & 0x1F
-			b[2] = (src[1] >> 1) & 0x1F
-			b[1] = (src[1] >> 6) & 0x1F
+			val |= uint64(src[1]) << 24
 			fallthrough
 		case 1:
-			b[1] |= (src[0] << 2) & 0x1F
-			b[0] = src[0] >> 3
+			val |= uint64(src[0]) << 32
 		}
 
 		// Encode 5-bit blocks using the base32 alphabet
-		size := len(dst)
+		size := uint(len(dst))
 		if size >= 8 {
 			size = 8
 		}
-		for i := 0; i < size; i++ {
-			dst[i] = enc.encode[b[i]&31]
+		for i := uint(0); i < size; i++ {
+			dst[i] = enc.encode[(val>>(35-5*i))&31]
 		}
 	}
 }
@@ -229,15 +219,40 @@ func (e CorruptInputError) Error() string {
 // additional data is an error. This method assumes that src has been
 // stripped of all supported whitespace ('\r' and '\n').
 func (enc *Encoding) decode(dst, src []byte) (n int, err error) {
-	// based on https://github.com/golang/go/blob/ba9e10889976025ee1d027db6b1cad383ec56de8/src/encoding/base32/base32.go#L277
-
 	// Lift the nil check outside of the loop.
 	_ = enc.decodeMap
 
-	dsti := 0
 	olen := len(src)
 
-	for len(src) > 0 {
+	// Decode in 8-byte chunks
+	for len(src) >= 8 {
+		var dbuf [8]byte
+		for j := 0; j < len(dbuf); j++ {
+			dbuf[j] = enc.decodeMap[src[j]]
+			if dbuf[j] == 0xff {
+				return n, CorruptInputError(olen - len(src) + j)
+			}
+		}
+		src = src[8:]
+
+		val := uint64(dbuf[0])<<35 |
+			uint64(dbuf[1])<<30 |
+			uint64(dbuf[2])<<25 |
+			uint64(dbuf[3])<<20 |
+			uint64(dbuf[4])<<15 |
+			uint64(dbuf[5])<<10 |
+			uint64(dbuf[6])<<5 |
+			uint64(dbuf[7])
+		dst[0] = byte(val >> 32)
+		dst[1] = byte(val >> 24)
+		dst[2] = byte(val >> 16)
+		dst[3] = byte(val >> 8)
+		dst[4] = byte(val)
+		n += 5
+		dst = dst[5:]
+	}
+
+	if len(src) > 0 {
 		// Decode quantum using the base32 alphabet
 		var dbuf [8]byte
 		dlen := len(src)
@@ -258,28 +273,27 @@ func (enc *Encoding) decode(dst, src []byte) (n int, err error) {
 		// quantum
 		switch dlen {
 		case 8:
-			dst[dsti+4] = dbuf[6]<<5 | dbuf[7]
+			dst[4] = dbuf[6]<<5 | dbuf[7]
 			n++
 			fallthrough
 		case 7:
-			dst[dsti+3] = dbuf[4]<<7 | dbuf[5]<<2 | dbuf[6]>>3
+			dst[3] = dbuf[4]<<7 | dbuf[5]<<2 | dbuf[6]>>3
 			n++
 			fallthrough
 		case 6, 5:
 			// dbuf[5] might be padding
-			dst[dsti+2] = dbuf[3]<<4 | dbuf[4]>>1
+			dst[2] = dbuf[3]<<4 | dbuf[4]>>1
 			n++
 			fallthrough
 		case 4:
-			dst[dsti+1] = dbuf[1]<<6 | dbuf[2]<<1 | dbuf[3]>>4
+			dst[1] = dbuf[1]<<6 | dbuf[2]<<1 | dbuf[3]>>4
 			n++
 			fallthrough
 		case 3, 2:
 			// dbuf[2] might be padding
-			dst[dsti+0] = dbuf[0]<<3 | dbuf[1]>>2
+			dst[0] = dbuf[0]<<3 | dbuf[1]>>2
 			n++
 		}
-		dsti += 5
 	}
 	return n, nil
 }