@shogo82148 2023-07-16
old:

```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/go-clockwork-base32
BenchmarkEncode-10            	  132634	      9042 ns/op	 905.95 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  132016	      9075 ns/op	 902.66 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  132789	      9015 ns/op	 908.72 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  132831	      8997 ns/op	 910.50 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  133093	      9000 ns/op	 910.22 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  132804	      9016 ns/op	 908.58 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  132735	      9006 ns/op	 909.57 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  132788	      9037 ns/op	 906.53 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  129798	      9058 ns/op	 904.44 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  132490	      9473 ns/op	 864.74 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncodeToString-10    	  104853	     10616 ns/op	 771.70 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  105292	     10739 ns/op	 762.81 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  115200	     10416 ns/op	 786.51 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  112269	     10806 ns/op	 758.10 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  108937	     11339 ns/op	 722.45 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  112124	     10589 ns/op	 773.60 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  113479	     10837 ns/op	 755.96 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  103411	     10701 ns/op	 765.55 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  113186	     10557 ns/op	 775.96 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  113514	     10474 ns/op	 782.16 MB/s	   27136 B/op	       2 allocs/op
BenchmarkDecode-10            	   80308	     14983 ns/op	 874.83 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	   80179	     15050 ns/op	 870.99 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	   80151	     14954 ns/op	 876.54 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	   80180	     14980 ns/op	 875.03 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	   80246	     14898 ns/op	 879.88 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	   80425	     14933 ns/op	 877.80 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	   80416	     14899 ns/op	 879.82 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	   80464	     14912 ns/op	 879.05 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	   80348	     14940 ns/op	 877.37 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	   80426	     14919 ns/op	 878.58 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecodeString-10      	   76269	     15694 ns/op	 835.24 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	   76233	     15672 ns/op	 836.38 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	   76524	     15672 ns/op	 836.40 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	   76339	     15701 ns/op	 834.87 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	   76239	     15822 ns/op	 828.48 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	   75014	     15933 ns/op	 822.71 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	   74624	     16017 ns/op	 818.37 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	   76150	     15850 ns/op	 827.02 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	   75223	     15794 ns/op	 829.96 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	   76492	     15709 ns/op	 834.43 MB/s	   13568 B/op	       1 allocs/op
PASS
ok  	github.com/shogo82148/go-clockwork-base32	54.390s
```

new:

```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/go-clockwork-base32
BenchmarkEncode-10            	  246968	      4848 ns/op	1689.86 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  247040	      4846 ns/op	1690.35 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  246628	      4852 ns/op	1688.39 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  246366	      4843 ns/op	1691.67 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  247394	      4854 ns/op	1687.69 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  246835	      4848 ns/op	1689.75 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  246885	      4848 ns/op	1689.84 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  246246	      4883 ns/op	1677.81 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  246727	      4844 ns/op	1691.30 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncode-10            	  246202	      4852 ns/op	1688.53 MB/s	       0 B/op	       0 allocs/op
BenchmarkEncodeToString-10    	  191976	      6194 ns/op	1322.60 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  191071	      6226 ns/op	1315.67 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  192124	      6246 ns/op	1311.57 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  191251	      6271 ns/op	1306.32 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  189628	      6249 ns/op	1310.96 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  188312	      6263 ns/op	1307.95 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  191104	      6299 ns/op	1300.56 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  190520	      6268 ns/op	1306.94 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  190447	      6266 ns/op	1307.36 MB/s	   27136 B/op	       2 allocs/op
BenchmarkEncodeToString-10    	  190818	      6272 ns/op	1306.20 MB/s	   27136 B/op	       2 allocs/op
BenchmarkDecode-10            	  135819	      8803 ns/op	1488.98 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	  134106	      8820 ns/op	1486.14 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	  135822	      8806 ns/op	1488.55 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	  135927	      8808 ns/op	1488.11 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	  135746	      8817 ns/op	1486.61 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	  136168	      8834 ns/op	1483.83 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	  135708	      8812 ns/op	1487.58 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	  135944	      8805 ns/op	1488.63 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	  135900	      8811 ns/op	1487.73 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecode-10            	  135793	      8812 ns/op	1487.55 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecodeString-10      	  124171	      9596 ns/op	1365.94 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	  123895	      9589 ns/op	1367.05 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	  124215	      9588 ns/op	1367.08 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	  124195	      9586 ns/op	1367.39 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	  124875	      9579 ns/op	1368.43 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	  124902	      9577 ns/op	1368.70 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	  124705	      9591 ns/op	1366.64 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	  124062	      9769 ns/op	1341.83 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	  116397	      9737 ns/op	1346.22 MB/s	   13568 B/op	       1 allocs/op
BenchmarkDecodeString-10      	  121197	     10270 ns/op	1276.39 MB/s	   13568 B/op	       1 allocs/op
PASS
ok  	github.com/shogo82148/go-clockwork-base32	51.979s
```